add: ENVによる役職一覧の切り替えを実装
